ventureSense/requirements.txt
websockets
google-generativeai
google-cloud-aiplatform
python-dotenv
google-adk
uvicorn
fastapi